CHROMIUM: drm/panel-simple: Support for delays between GPIO & regulator
Some panels datasheets may specify a delay between the enable GPIO and
the regulator. Support this in panel-simple.

This is a CHROMIUM patch because we're still working on how to best
deal with the panel that needs this (ATNA33XC20) upstream.

BUG=b:180993541, b:181690025, b:181690025
TEST=emerge-strongbad chromeos-kernel-5_4

Change-Id: Ie8cf2ccfdf11ff43ae17bfc6057b9dc75c6dbb61
Signed-off-by: yangcong <yangcong5@huaqin.corp-partner.google.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2847952
Reviewed-by: Sean Paul <seanpaul@chromium.org>
(cherry picked from commit 01591ea61bf42d68245a5797d0db865e27fedc55)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2949259