FROMGIT: drm/i915/stolen: actually mark as contiguous
Stolen memory is always allocated as physically contiguous pages, so
mark the object flags as such. It looks like the flags were previously
just ignored so this had no effect. In the future we might to add the
proper plumbing for passing the flags all over the way down from the
caller, but for now we don't have a use for that.

Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210421104658.304142-4-matthew.auld@intel.com
(cherry picked from commit ae7f0dc105c4e5248e6e2e2299e52d9a289e1215
 git://anongit.freedesktop.org/drm/drm drm-next)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:187300590,b:190020996
TEST=Tested on volteer/zork/trogdor, build tested on the rest

Change-Id: Iff12b167627861af4e4a173090fa975ed55fa8aa