FROMLIST: media: mtk-vcodec: vdec: move stateful ops into their own file
We are planning to add support for stateless decoders to this driver.
Part of the driver will be shared between stateful and stateless
codecs, but a few ops need to be specialized for both. Extract the
stateful part of the driver and move it into its own file, accessible
through ops that the common driver parts can call.

This patch only moves code around and introduces a set of abstractions ;
the behavior of the driver should not be changed in any way.

Signed-off-by: Yunfei Dong <yunfei.dong@mediatek.com>
[acourbot: refactor, cleanup and split]
Co-developed-by: Alexandre Courbot <acourbot@chromium.org>
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

(am from https://patchwork.kernel.org/project/linux-mediatek/patch/20210427111526.1772293-2-acourbot@chromium.org/)

BUG=b:187685837
TEST=boot krane

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I12375c79af8b5a6a8c2e609ec0894fea32eb754a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2883008
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>