CHROMIUM: anx7625: Don't assume that internal display is always connected.
When factory is building with PCB only, the internal display is not
attached, but currently the driver assumes that it is always attached.
This makes Chrome thinks there's an inexistent internal display, and put
the external display at extended mode, which makes it not usable.

Actually detect the internal display too in this case.

BUG=b:183076848
TEST=manually, disconnect the internal display and check

Change-Id: I06ab1699172aed0cc5bc441b8a79c509b145cc28
Signed-off-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2870414
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
(cherry picked from commit fd084a99ca2960f141d180dd4d8968ed09555278)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2916433