CHROMIUM: mac80211: extend protection against mixed key and fragment cache attacks
For some chips/drivers, e.g., QCA6174 with ath10k, the decryption is
done by the hardware, and the Protected bit in the Frame Control field
is cleared in the lower level driver before the frame is passed to
mac80211. In such cases, the condition for ieee80211_has_protected() is
not met in ieee80211_rx_h_defragment() of mac80211 and the new security
validation steps are not executed.

Extend mac80211 to cover the case where the Protected bit has been
cleared, but the frame is indicated as having been decrypted by the
hardware. This extends protection against mixed key and fragment cache
attack for additional drivers/chips. This fixes CVE-2020-24586 and
CVE-2020-24587 for such cases.

Tested-on: QCA6174 hw3.2 PCI WLAN.RM.4.4.1-00110-QCARMSWP-1

BUG=b:179418220
TEST=manual testing on M

Change-Id: I742ff7bffb9dabe0d9c3ca85f300932bc194263c
Signed-off-by: Wen Gong <wgong@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: vadiraj kakhandaki <vadirajk@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2922305
Reviewed-by: Julan Hsu <julanhsu@google.com>
Reviewed-by: Balaji Pothunoori <bpothuno%qti.qualcomm.com@gtempaccount.com>
Tested-by: ANKIT DARSI AMARNATHAGUPTHA <dankit@google.com>