UPSTREAM: vfio/iommu_type1: Use struct_size() for kzalloc()
Make use of the struct_size() helper instead of an open-coded version,
in order to avoid any potential type mistakes or integer overflows
that, in the worst scenario, could lead to heap overflows.

This code was detected with the help of Coccinelle and, audited and
fixed manually.

Signed-off-by: Gustavo A. R. Silva <gustavoars@kernel.org>
Message-Id: <20210513230155.GA217517@embeddedor>
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
(cherry picked from commit 78b238147e4d241bc1681d2559477c995f9dcb0a)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:187300590
TEST=Tested on volteer/zork/trogdor, build tested on the rest

Change-Id: I793636d5f589d33839755d70ec6e2733d0982a13