BACKPORT: drm/tegra: Get ref for DP AUX channel, not its ddc adapter
While we're taking a reference of the DDC adapter for a DP AUX channel in
tegra_sor_probe() because we're going to be using that adapter with the
SOR, now that we've moved where AUX registration happens the actual device
structure for the DDC adapter isn't initialized yet. Which means that we
can't really take a reference from it to try to keep it around anymore.

This should be fine though, because we can just take a reference of its
parent instead.

v2:
* Avoid calling i2c_put_adapter() in tegra_output_remove() for eDP/DP cases

Signed-off-by: Lyude Paul <lyude@redhat.com>
Fixes: 39c17ae60ea9 ("drm/tegra: Don't register DP AUX channels before connectors")
Cc: Lyude Paul <lyude@redhat.com>
Cc: Thierry Reding <thierry.reding@gmail.com>
Cc: Jonathan Hunter <jonathanh@nvidia.com>
Cc: dri-devel@lists.freedesktop.org
Cc: linux-tegra@vger.kernel.org
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 1d15a10395e5a036f571ac727f202f9572e255f9)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Downstream changes:
-Upstream merge conflict with 1bbcde403c20 ("UPSTREAM: drm/tegra: Get
 ref for DP AUX channel, not its ddc adapter") and 626d1b56c962
 ("UPSTREAM: drm/tegra: sor: Fix AUX device reference leak")

BUG=b:187300590
TEST=Tested with single/multiple displays, suspend/resume, hotplugs on volteer/trogdor/dedede/zork, build tested on amd64/arm64-generic

Change-Id: I4c344c8c52e18e5537c88d6495f8cd37f7e4a53e