FROMGIT: ALSA: intel8x0: Don't update period unless prepared
The interrupt handler of intel8x0 calls snd_intel8x0_update() whenever
the hardware sets the corresponding status bit for each stream.  This
works fine for most cases as long as the hardware behaves properly.
But when the hardware gives a wrong bit set, this leads to a zero-
division Oops, and reportedly, this seems what happened on a VM.

For fixing the crash, this patch adds a internal flag indicating that
the stream is ready to be updated, and check it (as well as the flag
being in suspended) to ignore such spurious update.

Cc: <stable@vger.kernel.org>
Reported-and-tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Link: https://lore.kernel.org/r/s5h5yzi7uh0.wl-tiwai@suse.de
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(cherry picked from commit c1f0616124c455c5c762b6f123e40bba5df759e6
 https://git.kernel.org/pub/scm/linux/kernel/git/tiwai/sound.git for-linus)

[5.4-arcvm: picked from 5.4]

BUG=b:188009314
TEST=tested that VM suspend/suspend doesn't crash the VM on hatch-arc-r

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Change-Id: Idfe3d18158114f3d8981b89147b97bba036ad559
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2901899
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
Commit-Queue: Suleiman Souhlal <suleiman@chromium.org>