UPSTREAM: drm/amdgpu/display: make backlight setting failure messages debug
Avoid spamming the log.  The backlight controller on DCN chips
gets powered down when the display is off, so if you attempt to
set the backlight level when the display is off, you'll get this
message.  This isn't a problem as we cache the requested backlight
level if it's adjusted when the display is off and set it again
during modeset.

Reviewed-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Cc: nicholas.choi@amd.com
Cc: harry.wentland@amd.com
(cherry picked from commit cd11b58ccef87c59918daa7eb0202fb308aa1350)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:187300590
TEST=Tested with single/multiple displays, suspend/resume, hotplugs on volteer/trogdor/dedede/zork, build tested on amd64/arm64-generic

Change-Id: Iab9864281b6d8accdb49bdfe82ccdfd1c071d674