FROMGIT: drm/i915/display: relax 2big checking around initial fb
The kernel prefers enabling fbc over the initial fb, since this leads to
actual runtime power savings, so if the initial fb is deemed too big
using some heuristic, then we simply skip allocating stolen for it.
However if the kernel is not configured with fbcon then it should be
possible to relax this, since unlike with fbcon the display server
shouldn't preserve it when later replacing it, and so we should be able
to re-use the stolen memory for fbc and friends. This patch is reported
to fix some flicker seen during boot splash on some devices.

v2: s/FRAMEBUFFER_CONSOLE/CONFIG_FRAMEBUFFER_CONSOLE

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Lee Shawn C <shawn.c.lee@intel.com>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: Daniel Vetter <daniel.vetter@intel.com>
Reviewed-by: Imre Deak <imre.deak@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210526124901.245689-1-matthew.auld@intel.com
(cherry picked from commit 40e40e63f7be028ff8dc2a2dcef4a812539e86cc
 git://anongit.freedesktop.org/drm/drm drm-next)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:187300590,b:190020996
TEST=Tested on volteer/zork/trogdor, build tested on the rest

Change-Id: I17c5c2143981d3aec256bbe588f08481228bd4fc