FROMGIT: drm/i915/adl_p: Same slices mask is not same Dbuf state
We currently treat same slice mask as a same DBuf state and skip
updating the Dbuf slices, if we detect that.
This is wrong as if we have a multi to single pipe change or
vice versa, that would be treated as a same Dbuf state and thus
no changes required, so we don't get Mbus updated, causing issues.
Solution: check also mbus_join, in addition to slices mask.

Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
Signed-off-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210527110106.21434-1-stanislav.lisovskiy@intel.com
(cherry picked from commit 0d6695b112762aa7aad28c46e65561389b6f50d6
 git://anongit.freedesktop.org/drm/drm drm-next)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:187300590,b:190020996
TEST=Tested on volteer/zork/trogdor, build tested on the rest

Change-Id: Ica78eba7ee2dfb42b93b2a663d4f1b133d3fd1d1