CHROMIUM: config: Turn off AMBA DMA driver on qcom
We don't use this driver on qcom flavours, but it is enabled in the
common arm64 config. Move it to the other device specific configs, i.e.
mediatek and rockchip, so that the AMBA DMA driver is only enabled
there.

This is manually crafted by removing the config from the common.config
file and adding it to the end of the other configs and then running

 ./chromeos/scripts/kernelconfig -F olddefconfig

BUG=b:189484858
TEST=tast run $TROGDOR_DUT kernel.Bloat; doesn't show this driver

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I1b320927f48d52c3a3c9d5db9b064b139ef3245c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2984070
Reviewed-by: Douglas Anderson <dianders@chromium.org>