CHROMIUM: ASoC: SOF: Intel: hda-dai: configure SSP in prepare stage
If a system suspend happens (e.g. with rtcwake in tests) while
playback/capture is on-going, the hw_params step is not invoked on
resume.

Other dais such as HDA and ALH/SoundWire use both the .hw_params and
the .prepare stage, but in the latter case use a state variable to
detect the difference between underflows and resume operations due to
restrictions on programming sequences.

For the SSP it's fine to only do the configuration in the .prepare
step.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>

BUG=b:188623237, b:192632626
TEST=ODM tests audio function on Keeby

Signed-off-by: Brent Lu <brent.lu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3060625
Tested-by: Ajye Huang <ajye_huang@compal.corp-partner.google.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Change-Id: Ic12d76316e18af9bdad153498e87646bd8ef40ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3187393
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>
Commit-Queue: Shou-Chieh Hsu <shouchieh@chromium.org>