CHROMIUM: drm/i915/lspcon: 1000ms LSPCON timeout
500ms was still not enough time for the Parade PS175 LSPCON adapter on
Ambassador to settle. This results in blank screens on boot. Factory
testing on larger sample sizes of devices resulted in a 2% failure rate.

We are increasing the timeout to unblock PVT production while we
continue to work on root-causing why the original 400ms is not enough
time to settle. The factory did not report any issues when using a build
that uses a 1000ms timeout, hence this change.

BUG=b:188035814
TEST=build kernel; deploy to device; reboot 100x

Change-Id: I03d0f8d4e8bbfe8bead4458fa35768d9b4e51393
Signed-off-by: Joe Tessler <jrt@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2945760
Tested-by: Joe Tessler <jrt@chromium.org>
Auto-Submit: Joe Tessler <jrt@chromium.org>
Reviewed-by: Matthew Ziegelbaum <ziegs@chromium.org>
Commit-Queue: Matthew Ziegelbaum <ziegs@chromium.org>